src/function_app/requirements.txt
azure-functions
pillow>=10.0.0
azure-identity>=1.12.0